+++ trueseeing/core/model/issue.py
