+++ trueseeing/sig/fingerprint.py
@@ -5,10 +5,10 @@ import io
 import os
 import re
 
-from trueseeing.core.code.model import InvocationPattern
-from trueseeing.core.flow.data import DataFlows
-from trueseeing.signature.base import Detector
-from trueseeing.core.issue import Issue
+from trueseeing.core.model.code import InvocationPattern
+from trueseeing.core.analysis.flow import DataFlows
+from trueseeing.core.model.sig import Detector
+from trueseeing.core.model.issue import Issue
 
 if TYPE_CHECKING:
   from typing import Iterable, Optional, List, Dict, Any, Set
