+++ trueseeing/sig/privacy.py
@@ -3,15 +3,15 @@ from typing import TYPE_CHECKING
 
 import re
 
-from trueseeing.core.code.model import InvocationPattern
-from trueseeing.core.flow.data import DataFlows
-from trueseeing.signature.base import Detector
-from trueseeing.core.issue import Issue
+from trueseeing.core.model.code import InvocationPattern
+from trueseeing.core.analysis.flow import DataFlows
+from trueseeing.core.model.sig import Detector
+from trueseeing.core.model.issue import Issue
 
 if TYPE_CHECKING:
   from typing import Optional
   from trueseeing.core.store import Store
-  from trueseeing.core.code.model import Op
+  from trueseeing.core.model.code import Op
 
 class PrivacyDeviceIdDetector(Detector):
   option = 'privacy-device-id'
